requirements.txt
flask
requests
dnspython
python-whois